requirements.txt
torch>=2.1.0
torchvision>=0.16.0
streamlit>=1.30.0
pillow>=10.0.0
transformers>=4.40.0
skorch>=1.3.0
pandas>=2.0.0
numpy>=1.26.0
seaborn>=0.13.0
matplotlib>=3.8.0
tqdm>=4.66.0
scikit-learn>=1.4.0
scipy>=1.11.0